breakup spec by journable
